% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BGHRsample.R
\docType{data}
\name{BGHRsample}
\alias{BGHRsample}
\title{Information for each electrofishing sample from Big Hill Reservoir, KS, 2014.}
\format{A data frame with 20 observations on the following 4 variables.
 \describe{
   \item{UID}{Unique sample identification number}
   \item{date}{Data sample was collected}
   \item{loc}{Location code for where the sample was collected}
   \item{effort}{Effort (minutes) expended for the sample}
 }}
\source{
Data obtained directly from Ben Neely.  Used in the \href{http://derekogle.com/IFAR/}{Introductory Fisheries Analyses with R} book.
}
\description{
Information for each electrofishing sample from Big Hill Reservoir, KS, in May, 2014.
}
\section{Topic(s)}{

 \itemize{
   \item Data Manipulation
 }
}
\examples{
data(BGHRsample)
str(BGHRsample)
head(BGHRsample)
}
\seealso{
See \code{\link{BGHRfish}} for individual fish collected in these samples.
}
\concept{
'Data Manipulation'
}
\keyword{datasets}

