% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BluegillIL.R
\docType{data}
\name{BluegillIL}
\alias{BluegillIL}
\title{Length-at-marking and recapture and time-at-large of Bluegill.}
\format{A data frame with 61 observations on the following 5 variables:
 \describe{
   \item{tag}{Unique Floy tag number.}
   \item{lenMark}{length (mm) at tagging.}
   \item{lenRecap}{Length (mm) at recapture.}
   \item{deltaLen}{Change in length (mm).}
   \item{deltaTime}{Time-at-large (yrs).} 
 }}
\source{
These unpublished data are from Derek H. Ogle, Northland College.  \bold{Do not use for other than educational purposes without permission from the source.}
}
\description{
Length-at-marking and recapture and time-at-large for Bluegill (\emph{Lepomis macrochirus}) originally captured in Spring 2010 from Inch Lake, Wisconsin.
}
\section{Topic(s)}{

 \itemize{
   \item Growth 
   \item von Bertalanffy 
   \item Fabens method 
 }
}
\examples{
data(BluegillIL)
str(BluegillIL)
head(BluegillIL)
plot((lenRecap-lenMark)~deltaTime,data=BluegillIL)

}
\concept{
Growth 'von Bertalanffy' Fabens
}
\keyword{datasets}

