\docType{data}
\name{BluntnoseIL1}
\alias{BluntnoseIL1}
\title{Lengths of Bluntnose Minnows (subsample) captured in Inch Lake}
\format{A data frame of 25 observations on the following 3 variables:
\describe{
\item{netID}{a netID factor (all 41).}
\item{species}{species name (all Bluntnose)}
\item{tl}{total length (inches to nearest 0.1)}
}}
\source{
Derek H. Ogle, personal collection
}
\description{
Total lengths of a subsample of 144 Bluntnose Minnows
(\emph{Pimephales notatus}) captured in Inch Lake, Bayfield
County, WI in May, 2007.
}
\section{Topic(s)}{
  \itemize{ \item Length frequency \item Length expansion }
}
\examples{
data(BluntnoseIL1)
str(BluntnoseIL1)
head(BluntnoseIL1)
}
\seealso{
\code{\link{InchLake1}}, \code{\link{InchLake2}}.
}
\keyword{datasets}

