% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrookTroutNC.R
\docType{data}
\name{BrookTroutNC}
\alias{BrookTroutNC}
\title{Stock and recruitment data for Brook Trout from Ball Creek, NC.}
\format{A data frame with 10 observations on the following 2 variables.
 \describe{
   \item{adult}{a numeric vector giving autumn adult density (number per square meter).}
   \item{yoy}{a numeric vector giving autumn YOY density (number per square meter) in following year.}
 }}
\source{
Data extracted approximately (with digitize()) from figure 5 in Grossman, G.D., R.E. Ratajczak, C.M. Wagner, and J.T. Petty.  2010. Dynamics and regulation of the southern brook trout (\emph{Salvelinus fontinalis}) population in an Appalachian stream.  Freshwater Biology 55:1494-1508.
}
\description{
Stock and recruitment data for Brook Trout (\emph{Salvelinus fontinalis}) from Ball Creek, NC.
}
\note{
The authors fit a linear model to the stock-recruit relationship.
}
\section{Topic(s)}{

 \itemize{
   \item Stock-recruit
 }
}
\examples{
data(BrookTroutNC)
str(BrookTroutNC)
head(BrookTroutNC)
plot(adult~yoy,data=BrookTroutNC)
}
\concept{
'Stock-Recruit'
}
\keyword{datasets}

