% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FSAdata.R
\docType{package}
\name{FSAdata}
\alias{FSAdata}
\alias{FSAdata-package}
\title{Data to support the FSA package.}
\description{
This package contains data to support the FSA package.
}
\details{
This package contains additional data files that can be used for common fisheries stock assessment methods described in the \pkg{FSA} package and on the \href{https://fishr.wordpress.com/}{fishR webpage}.

The help files for these datasets are embedded with topics that can be searched to find data files that can be analyzed with those topics.  For example, use the following commands to find data files for the corresponding topics.

\tabular{ll}{
 \code{help.search("Length-Weight",package=c("FSAdata","FSA"))} \tab Length-weight model data.\cr
 \code{help.search("Growth",package=c("FSAdata","FSA"))} \tab Growth model data.\cr
 \code{help.search("Bertalanffy",package=c("FSAdata","FSA"))} \tab von Bertalanffy growth model data.\cr
 \code{help.search("Stock-Recruit",package=c("FSAdata","FSA"))} \tab Stock-recruitment data.\cr
 \code{help.search("Recruitment",package=c("FSAdata","FSA"))} \tab Recruitment time-series data.\cr
 \code{help.search("Abundance",package=c("FSAdata","FSA"))} \tab Data for abundance estimates.\cr
 \code{help.search("Mark-recapture",package=c("FSAdata","FSA"))} \tab Mark-recapture data.\cr
 \code{help.search("Capture history",package=c("FSAdata","FSA"))} \tab Capture history mark-recapture (compare to summarized data) data.\cr
 \code{help.search("Schnabel",package=c("FSAdata","FSA"))} \tab Schnabel mark-recapture (closed population, multiple samples).\cr
 \code{help.search("Depletion",package=c("FSAdata","FSA"))} \tab Depletion (Leslie, DeLury) methods for estimating abundance.\cr
 \code{help.search("Removal",package=c("FSAdata","FSA"))} \tab Removal (K-pass) methods for estimating abundance.\cr
 \code{help.search("Age Comparison",package=c("FSAdata","FSA"))} \tab Ageing (error, precision, or comparison).\cr
 \code{help.search("Catch curve",package=c("FSAdata","FSA"))} \tab Catch curve.\cr
 \code{help.search("Mortality",package=c("FSAdata","FSA"))} \tab Data for mortality estimation.\cr
 \code{help.search("Maturity",package=c("FSAdata","FSA"))} \tab Maturity data.\cr
 \code{help.search("Size Structure",package=c("FSAdata","FSA"))} \tab Size structure data.\cr
 \code{help.search("Age-Length Key",package=c("FSAdata","FSA"))} \tab Age-Length Key data.\cr
}

Several additional data sets can be found in the FSA package (development version found at \url{https://fishr.wordpress.com/fsa/}).
}

