% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LakeTroutEggs.R
\docType{data}
\name{LakeTroutEggs}
\alias{LakeTroutEggs}
\title{Length and egg deposition of Lake Superior Lake Trout.}
\format{A data frame of 101 observations on the following 2 variables:
 \describe{
   \item{tl}{Total length (mm) of an individual fish.}
   \item{eggs}{Estimated number of eggs.}
 }}
\source{
Schram, S.T. 1993.  Fecundity and egg deposition of a wild Lake Superior Lake Trout stock.  Wisconsin Department of Natural Resources, Fisheries Management Report no. 149.
}
\description{
Length and egg deposition of Lake Superior Lake Trout (\emph{Salvelinus namaycush}).
}
\examples{
data(LakeTroutEggs)
str(LakeTroutEggs)
head(LakeTroutEggs)
plot(eggs~tl,data=LakeTroutEggs)
}
\keyword{datasets}

