% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morwong2.R
\docType{data}
\name{Morwong2}
\alias{Morwong2}
\title{Ages of Morwong assigned from otoliths by Reader B at two times.}
\format{
A data frame with 136 observations on the following 2 variables.
 \describe{
   \item{first}{Ages assigned on the first reading}
   \item{second}{Ages assigned on the second reading}
 }
}
\source{
From Table 6 of Restall, J.E., and K. Krusic-Golub. 2004. Development of jackass morwong age-length keys for 2000-2002. Final report to Australian Fisheries Management Authority. 13 pp. Primary Industries Research Victoria, Queenscliff. \href{https://raw.githubusercontent.com/fishR-Core-Team/FSAdata/main/data-raw/Morwong2.csv}{CSV file}
}
\description{
Ages assigned at two different times by Reader B to the otoliths of Jackass Morwong (\emph{Nemadactylus macropterus}).
}
\section{Topic(s)}{

 \itemize{
   \item Age Comparison 
   \item Age Precision 
   \item Age Bias
   \item Ageing Error
 }
}

\examples{
data(Morwong2)
str(Morwong2)
head(Morwong2)
plot(second~first,data=Morwong2)
xtabs(~first+second,data=Morwong2)

}
\concept{Age Bias}
\concept{Age Comparison}
\concept{Age Precision}
\keyword{datasets}
