% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morwong3.R
\docType{data}
\name{Morwong3}
\alias{Morwong3}
\title{Ages of Morwong assigned from otoliths by two readers.}
\format{A data frame with 58 paired observations on the following 2 variables.
 \describe{
   \item{readerA}{Ages assigned by Reader A}
   \item{readerB}{Ages assigned by Reader B}
 }}
\source{
From Table 7  of Restall, J.E., and K. Krusic-Golub. 2004. Development of jackass morwong age-length keys for 2000-2002. Final report to Australian Fisheries Management Authority. 13 pp. Primary Industries Research Victoria, Queenscliff. Was (is?) available at http://web-test.afma.gov.au/wp-content/uploads/2010/07/r03_1724b.pdf.
}
\description{
Ages assigned by two different readers to the otoliths of Jackass Morwong (\emph{Nemadactylus macropterus}).
}
\section{Topic(s)}{

 \itemize{
   \item Age Comparison
   \item Age Precision 
   \item Age Bias
   \item Ageing Error
 }
}

\examples{
data(Morwong3)
str(Morwong3)
head(Morwong3)
plot(readerB~readerA,data=Morwong3)
with(Morwong3,table(readerA,readerB))

}
\concept{Age Bias}
\concept{Age Comparison}
\concept{Age Precision}
\keyword{datasets}
