% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Morwong4a.R
\docType{data}
\name{Morwong4a}
\alias{Morwong4a}
\title{Age (subsample) and length (all fish) of Morwong from Morwong4.}
\format{A data frame with 392 observations on the following 2 variables.
 \describe{
   \item{fl}{Fork lengths (cm).}
   \item{age}{Assigned ages.}
 }}
\description{
A total of 104 fish dispersed proportionately over 1-cm fork length intervals from the \code{Morwong4} data frame was obtained for age assignment.  The remaining fish in the file were only measured for length (i.e., the ages were deleted).  This data file can be used to demonstrate the use of age-length keys.
}
\section{Topic(s)}{

 \itemize{
   \item Age-length key
 }
}
\examples{
data(Morwong4a)
str(Morwong4a)
head(Morwong4a)

## extract aged sample
m4a.aged <- subset(Morwong4a,!is.na(age))
str(m4a.aged)

## extract length sample
m4a.length <- subset(Morwong4a,is.na(age))
str(m4a.length)
}
\seealso{
\code{\link{Morwong4}}.
}
\concept{
'Age-Length Key'
}
\keyword{datasets}

