% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PikeIL.R
\docType{data}
\name{PikeIL}
\alias{PikeIL}
\title{Catch and effort data for Northern Pike from Island Lake, NB.}
\format{A data frame with 9 observations on the following 3 variables:
 \describe{
   \item{date}{Capture date.} 
   \item{nets}{Number of fyke nets fished on that date.} 
   \item{catch}{Number of captured and removed pike.} 
 }}
\source{
From Table 1 of Peterson, J., M. Taylor, and A. Hanson. 1980. Leslie population estimate for a large lake. Transactions of the American Fisheries Society, 109:329-331.
}
\description{
Catch and effort data for nine removal events of Northern Pike (\emph{Esox lucius}) from Island Lake, NB.
}
\section{Topic(s)}{

 \itemize{
   \item Depletion methods 
   \item Leslie method
   \item DeLury method 
   \item Population size 
   \item Abundance 
   \item Catchability
 }
}

\examples{
data(PikeIL)
str(PikeIL)
head(PikeIL)

}
\concept{Abundance}
\concept{Catchability}
\concept{DeLury}
\concept{Depletion}
\concept{Leslie}
\concept{Population Size}
\keyword{datasets}
