\docType{data}
\name{SalmonADP}
\alias{SalmonADP}
\title{Catch in removal events of salmon parr.}
\format{A data frame with 5 observations on the following 2 variables:
\describe{
\item{catch}{Removal event.}
\item{caught}{Number of fish captured and removed.}
}}
\source{
From Table 2 in Cowx, I.G.  1983.  Review of the methods
for estimating fish population size from survey removal
data.  Fisheries Management, 14:67-82.
}
\description{
Catch in removal events of salmon parr from the Afon Dulas
at Pentre, 19Jun79.
}
\examples{
data(SalmonADP)
str(SalmonADP)
SalmonADP
}
\seealso{
\link{TroutADP}
}
\keyword{datasets}

