% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StripedBass1.R
\docType{data}
\name{StripedBass1}
\alias{StripedBass1}
\title{Ages of Striped Bass assigned from scales and otoliths.}
\format{A data frame of 343 observations on the following 2 variables:
 \describe{
   \item{ageO}{Age assigned from examinaton of otoliths.}
   \item{ageS}{Age assigned from examination of scales.}
 }}
\source{
Extracted from Figure 6 in Chapter 10 (Striped Bass) of the VMRC Final Report on Finfish Ageing, 2000 by the \href{http://ww2.odu.edu/sci/cqfe/}{Center for Quantitative Fisheries Ecology} at Old Dominion University.
}
\description{
Ages of Striped Bass (\emph{Morone saxatilis}) assessed from heat-pressed
scales and cracked-and-burnt otoliths.
}
\section{Topic(s)}{

 \itemize{
   \item Age comparison
   \item Age precision
   \item Age Bias
   \item Ageing error
 }
}
\examples{
data(StripedBass1)
str(StripedBass1)
head(StripedBass1)
plot(ageS~ageO,data=StripedBass1)
xtabs(~ageO+ageS,data=StripedBass1)
}
\concept{
Age Precision Bias
}
\keyword{datasets}

