% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TroutADP.R
\docType{data}
\name{TroutADP}
\alias{TroutADP}
\title{Catches in removal events of trout.}
\format{
A data frame with 5 observations on the following 2 variables:
 \describe{
   \item{catch}{Removal event.} 
   \item{caught}{Number of fish captured and removed.} 
 }
}
\source{
From Table 1 in Cowx, I.G. 1983. Review of the methods for estimating fish population size from survey removal data. Fisheries Management 14:67-82. \href{https://raw.githubusercontent.com/fishR-Core-Team/FSAdata/main/data-raw/TroutADP.csv}{CSV file}
}
\description{
Catches of trout on five successive removal events from the Afon Dulas at Pentre, 19Jun79.
}
\section{Topic(s)}{

 \itemize{
   \item Population size
   \item Abundance
   \item Removal
 }
}

\examples{
data(TroutADP)
str(TroutADP)
TroutADP

}
\seealso{
\link{SalmonADP}
}
\concept{Abundance}
\concept{Population Size}
\concept{Removal}
\keyword{datasets}
