% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{WalleyeConsumption}
\alias{WalleyeConsumption}
\title{Consumption of prey by Walleye.}
\format{A data frame of 23 observations on the following 2 variables:
\describe{
\item{PreyDensity}{Density of prey (mg per g per day).}
\item{FoodConsump}{Food consumption by predator (mg per cubic meter)}
}}
\source{
Madenjian, C.P., and S.R. Carpenter. 1991. Individual-based model for
growth of young-of-the-year walleye: A piece of the recruitment puzzle.
Ecological Applications.  1:268-279.
}
\description{
Consumption of prey by Walleye at different prey densities.
}
\section{Topic(s)}{
 \itemize{
\item Nonlinear modeling
}
}
\examples{
data(WalleyeConsumption)
str(WalleyeConsumption)
head(WalleyeConsumption)
plot(FoodConsump~PreyDensity,data=WalleyeConsumption)
}
\keyword{datasets}

