\docType{data}
\name{WalleyeMN06b}
\alias{WalleyeMN06b}
\title{Summarized multiple mark-recapture data for Walleye from four lakes in
Northern Minnesota.}
\format{A data frame with 20 observations on the following 5 variables.
\describe{
\item{lake}{Studied lake.}
\item{date}{Capture date.}
\item{catch}{Total fish captured in each sample.}
\item{recap}{Marked fish captured in each sample.}
\item{retMark}{Marked fish returned to the population.}
}}
\source{
Data from appendix 2 in Borkholder, B.D., A.J. Edwards, and
C. Olson. 2007.  Spring adult and fall juvenile walleye
popluation surveys within the 1854 ceded territory of
Minnesota, 2006.  Fond du Lac Division of Resource
Management, Technical Report 41
}
\description{
Summary results of capture histories (number captured,
number of recaptured fish, and number of unmarked fish that
were marked) for Walleye (\emph{Sander vitreus}) collected
from four lakes in Northern Minnesota, USA.
}
\section{Topic(s)}{
  \itemize{ \item Population size \item Abundance \item
  Mark-recapture \item Schnabel method \item
  Schumacher-Eschmeyer method }
}
\examples{
data(WalleyeMN06b)
str(WalleyeMN06b)
head(WalleyeMN06b)
}
\keyword{datasets}

