% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BluefishAge.R
\docType{data}
\name{BluefishAge}
\alias{BluefishAge}
\title{Ages of Bluefish assigned from otoliths by two readers.}
\format{A data frame with 445 observations on the following 2 variables.
\describe{
  \item{r1}{Ages assigned by the first reader}
  \item{r2}{Ages assigned by the second reader}
}}
\source{
From Figure 2 in Chapter 3 (Bluefish) of the VMRC Final Report on Finfish Ageing, 2003 by the \href{http://www.odu.edu/sci/research/cqfe}{Center for Quantitative Fisheries Ecology} at Old Dominion University.
}
\description{
Ages assigned to Bluefish (\emph{Pomatomus saltatrix}) otoliths by two readers.
}
\section{Topic(s)}{
 \itemize{
  \item Age Comparison
  \item Age Precision
  \item Age Bias
  \item Ageing Error
 }
}
\examples{
data(BluefishAge)
str(BluefishAge)
head(BluefishAge)
plot(r1~r2,data=BluefishAge)
}
\concept{
Age Precision Bias 'Age Comparison'
}
\keyword{datasets}

