% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BluntnoseIL1.R
\docType{data}
\name{BluntnoseIL1}
\alias{BluntnoseIL1}
\title{Subampled lengths of Bluntnose Minnows from Inch Lake, WI.}
\format{A data frame of 25 observations on the following 3 variables:
  \describe{
    \item{netID}{a netID factor (all 41)}
    \item{species}{species name (all Bluntnose)}
    \item{tl}{total length (inches to nearest 0.1)}
  }}
\source{
Derek H. Ogle, personal collection
}
\description{
Total lengths for a subsample from 144 Bluntnose Minnows (\emph{Pimephales notatus}) from Inch Lake, WI in May, 2007.
}
\section{Topic(s)}{

  \itemize{
    \item Length Frequency
    \item Length Expansion
    \item Size Structure
    \item PSD
  }
}
\examples{
data(BluntnoseIL1)
str(BluntnoseIL1)
head(BluntnoseIL1)
}
\seealso{
\code{\link{InchLake1}}, \code{\link{InchLake2}}.
}
\concept{
'Length Frequency' 'Size Structure' PSD 'Length Expansion'
}
\keyword{datasets}

