% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Croaker1.R
\docType{data}
\name{Croaker1}
\alias{Croaker1}
\title{Ages of Atlantic Croaker assigned from otoliths by two readers.}
\format{A data frame of 317 observations on the following 2 variables:
 \describe{
   \item{reader1}{Age assigned by the first reader.}
   \item{reader2}{Age assigned by the second reader.}
 }}
\source{
Extracted from Figure 2 in Chapter 1 (Atlantic Croaker) of the VMRC Final Report on Finfish Ageing, 1999 by the \href{http://ww2.odu.edu/sci/cqfe/}{Center for Quantitative Fisheries Ecology} at Old Dominion University.
}
\description{
Otolith age of Atlantic croaker (\emph{Micropogonias undulatus}) from two readers.
}
\section{Topic(s)}{

 \itemize{
   \item Age comparison
   \item Age precision
   \item Age Bias
   \item Ageing error
 }
}
\examples{
data(Croaker1)
str(Croaker1)
head(Croaker1)
plot(reader2~reader1,data=Croaker1)
xtabs(~reader1+reader2,data=Croaker1)
}
\concept{
Age Precision Bias 'Age Comparison'
}
\keyword{datasets}

