% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMBassBL.R
\docType{data}
\name{LMBassBL}
\alias{LMBassBL}
\title{Lengths for Largemouth Bass from Boomer Lake, OK.}
\format{A data frame of 447 observations on the following 1 variable:
 \describe{
   \item{tl}{measured total length (cm)} 
 }}
\source{
From McNew, R.W. and R.C. Summerfelt.  1978.  Evaluation of a maximum-likelihood estimator for analysis of length-frequency distributions.  Transactions of the American Fisheries Society 107:730-736.  Data was simulated (uniform distribution of values within length bin) from summarized length frequencies in http://fishbase.org/.
}
\description{
Length measurements of Largemouth Bass (\emph{Micropterus salmoides}) from Boomer Lake, Oklahoma, in 1974.
}
\section{Topic(s)}{

 \itemize{
   \item Length Frequency
   \item Size Structure
   \item PSD
 }
}
\examples{
data(LMBassBL)
str(LMBassBL)
head(LMBassBL)
hist(LMBassBL$tl,main="")

}
\concept{
'Length Frequency' 'Size Structure' PSD
}
\keyword{datasets}

