% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PikeHL.R
\docType{data}
\name{PikeHL}
\alias{PikeHL}
\title{Capture histories (2 samples) of Northern Pike from Harding Lake.}
\format{
A data frame with 481 observations on the following 3 variables.
 \describe{
   \item{fish}{a numeric vector of unique fish identification numbers}
   \item{first}{a numeric vector of indicator variables for the first sample (1=captured)}
   \item{second}{a numeric vector of indicator variables for the second sample (1=captured)}
 }
}
\source{
Capture histories simulated from summarzed data in table 2 and text of Burkholder, A.  1991. Abundance and composition of northern pike, Harding Lake, 1990. Fishery Data Series 91-9, Alaksa Department of Fish and Game. Was (is?) from http://www.sf.adfg.state.ak.us/FedAidpdfs/Fds91-09.pdf
}
\description{
Capture histories for Northern Pike (\emph{Esox lucius}) captured from Harding Lake, 1990.
}
\note{
Only Northern Pike >449 mm were considered here.
}
\section{Topic(s)}{

 \itemize{
   \item Population Size
   \item Abundance
   \item Mark-Recapture
   \item Capture-Recapture
   \item Petersen
   \item Capture History
 }
}

\examples{
data(PikeHL)
str(PikeHL)
head(PikeHL)

}
\concept{Abundance}
\concept{Capture History}
\concept{Capture-Recapture}
\concept{Mark-Recapture}
\concept{Petersen}
\concept{Population Size}
\keyword{datasets}
