% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBTroutUNSP.R
\docType{data}
\name{RBTroutUNSP}
\alias{RBTroutUNSP}
\title{Capture histories (2 sample) of Rainbow Trout.}
\format{
A data frame with 173 observations on the following 2 variables:
 \describe{
   \item{first}{Indicator variable for the first sample (1=captured)}
   \item{second}{Indicator variable for the second sample (1=captured)}
 }
}
\source{
Simulated from Table 10 in of Warren, C.D., K.A. Frank, and F.E. Partridge. 2004. Federal Aid in Fish Restoration - 2000 job performance report. Idaho Department of Fish and Game.
}
\description{
Capture histories for Rainbow Trout (\emph{Oncorhynchus mykiss}) in Upper Niagara Springs pond.
}
\section{Topic(s)}{

 \itemize{
   \item Population Size
   \item Abundance
   \item Mark-Recapture
   \item Capture-Recapture 
   \item Petersen
   \item Capture History 
 }
}

\examples{
data(RBTroutUNSP)
str(RBTroutUNSP)
head(RBTroutUNSP)

}
\concept{Abundance}
\concept{Capture History}
\concept{Capture-Recapture}
\concept{Mark-Recapture}
\concept{Petersen}
\concept{Population Size}
\keyword{datasets}
