% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RWhitefishAI.R
\docType{data}
\name{RWhitefishAI}
\alias{RWhitefishAI}
\title{Ages and lengths of Round Whitefish.}
\format{A data frame with 995 observations on the following 2 variables.
  \describe{
    \item{tl}{Total length (mm).}
    \item{age}{Age (from scales).}
  }}
\source{
From Table 3 (a random tenths digit was added to the TL) in Bailey, M.M.  1963.  \href{http://fishbull.noaa.gov/63-1/bailey.pdf}{Age, growth, and maturity of round whitefish of the Apostle Islands and Isle Royale Regions, Lake Superior.}  Fishery Bulletin, 63:63-75.
}
\description{
Ages and total lengths of Round Whitefish (\emph{Prosopium cylindraecum}) collected from the Apostles Islands, Lake Superior.
}
\section{Topic(s)}{

  \itemize{
    \item Growth
    \item von Bertalanffy
  }
}
\examples{
data(RWhitefishAI)
str(RWhitefishAI)
head(RWhitefishAI)
plot(tl~age,data=RWhitefishAI)
}
\concept{
Growth 'von Bertalanffy'
}
\keyword{datasets}

