% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SalmonADP.R
\docType{data}
\name{SalmonADP}
\alias{SalmonADP}
\title{Catches in removal events of salmon parr.}
\format{A data frame with 5 observations on the following 2 variables:
\describe{
\item{catch}{Removal event.}
\item{caught}{Number of fish captured and removed.}
}}
\source{
From Table 2 in Cowx, I.G.  1983.  Review of the methods for estimating
fish population size from survey removal data.  Fisheries Management, 14:67-82.
}
\description{
Catches in removal events of salmon parr from the Afon Dulas at Pentre, 19Jun79.
}
\section{Topic(s)}{
 \itemize{
\item Population size
\item Abundance
\item Removal
}
}
\examples{
data(SalmonADP)
str(SalmonADP)
SalmonADP
}
\seealso{
\link{TroutADP}
}
\concept{
Abundance 'Population Size' Removal
}
\keyword{datasets}

