% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StripedBass2.R
\docType{data}
\name{StripedBass2}
\alias{StripedBass2}
\title{Ages and lengths of Atlantic Ocean Striped Bass.}
\format{A data frame of 1201 observations on the following 2 variables:
 \describe{
   \item{age}{Assigned ages (from scales).}
   \item{tl}{Measured total lengths (mm).}
 }}
\source{
From Table 1 in Chapter 10 (Striped Bass) of the VMRC Final Report on Finfish Ageing, 2003 by the \href{http://www.odu.edu/sci/research/cqfe}{Center for Quantitative Fisheries Ecology} at Old Dominion University.
}
\description{
Assigned ages (from otoliths) and measured total lengths for each of 1201 Striped Bass (\emph{Morone saxatilis}) from the Atlantic Ocean.
}
\section{Topic(s)}{

 \itemize{
   \item Growth
   \item von Bertalanffy
 }
}
\examples{
data(StripedBass2)
str(StripedBass2)
head(StripedBass2)
plot(tl~age,data=StripedBass2)
}
\seealso{
\code{\link{StripedBass3}}.
}
\concept{
Growth 'von Bertalanffy'
}
\keyword{datasets}

