% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StripedBass5.R
\docType{data}
\name{StripedBass5}
\alias{StripedBass5}
\title{Ages of Striped Bass assigned from otoliths by two readers.}
\format{A data frame of 458 observations on the following 2 variables:
 \describe{
   \item{reader1}{Ages assigned by the first reader.}
   \item{reader2}{Ages assigned by the second reader.}
 }}
\source{
Extracted from Figure 5 in Chapter 10 (Striped Bass) of the VMRC Final Report on Finfish Ageing, 2003 by the \href{http://ww2.odu.edu/sci/cqfe/}{Center for Quantitative Fisheries Ecology} at Old Dominion University.
}
\description{
Ages assigned by two different readers to the otoliths of Striped Bass (\emph{Morone saxatilis}).
}
\section{Topic(s)}{

 \itemize{
   \item Age comparison
   \item Age precision
   \item Age Bias
   \item Ageing error
 }
}
\examples{
data(StripedBass5)
str(StripedBass5)
head(StripedBass5)
plot(reader2~reader1,data=StripedBass5)
xtabs(~reader1+reader2,data=StripedBass5)
}
\seealso{
\code{\link{StripedBass4}} and \code{\link{StripedBass6}}.
}
\concept{
Age Precision Bias 'Age Comparison'
}
\keyword{datasets}

