% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StripedBass6.R
\docType{data}
\name{StripedBass6}
\alias{StripedBass6}
\title{Ages of Striped Bass assigned from scales and otoliths.}
\format{A data frame of 451 observations on the following 2 variables:
 \describe{
  \item{scale}{Ages assigned to scales}
  \item{otolith}{Ages assigned to otoliths}
 }}
\source{
From Figure 6 in Chapter 10 (Striped Bass) of the VMRC Final Report on Finfish Ageing, 2003 by the \href{http://ww2.odu.edu/sci/cqfe/}{Center for Quantitative Fisheries Ecology} at Old Dominion University.
}
\description{
Ages assigned to the scales and otoliths of Striped Bass (\emph{Morone saxatilis}).
}
\section{Topic(s)}{

\itemize{
  \item Age Comparison
  \item Age Precision
  \item Age Bias
  \item Ageing Error
 }
}
\examples{
data(StripedBass6)
str(StripedBass6)
head(StripedBass6)
plot(scale~otolith,data=StripedBass6)
xtabs(~otolith+scale,data=StripedBass6)
}
\seealso{
\code{\link{StripedBass4}} and \code{\link{StripedBass5}}.
}
\concept{
Age Precision Bias 'Age Comparison'
}
\keyword{datasets}

