% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TroutperchLM2.R
\docType{data}
\name{TroutperchLM2}
\alias{TroutperchLM2}
\title{Lengths for Troutperch from Lake Michigan, 1977.}
\format{
A data frame of 3385 observations on the following 1 variable:
 \describe{
   \item{fl}{fork length (mm)} 
 }
}
\source{
From Brandt, S.B., J.J. Magnuson, and L.B. Crowder. 1980. Thermal habitat partitioning by fishes in Lake Michigan. Canadian Journal of Fisheries and Aquatic Sciences. 37:1557-1564. Data was simulated (uniform distribution of values within length bin) from summarized length frequencies in http://fishbase.org/.
}
\description{
Troutperch (\emph{Percopsis omiscomaycus}) fork lengths from near Grand Haven, Lake Michigan, 1977.
}
\section{Topic(s)}{

 \itemize{
   \item Length Frequency 
   \item Size Structure
 }
}

\examples{
data(TroutperchLM2)
str(TroutperchLM2)
head(TroutperchLM2)
hist(TroutperchLM2$fl,main="")

}
\concept{Length Frequency}
\concept{Size Structure}
\keyword{datasets}
