% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TroutperchLM3.R
\docType{data}
\name{TroutperchLM3}
\alias{TroutperchLM3}
\title{Subsampled lengths of Troutperch from Lake Michigan, 1977.}
\format{A data frame of 300 observations on the following 1 variable:
 \describe{
   \item{fl}{Fork length (mm)}
 }}
\description{
Troutperch (\emph{Percopsis omiscomaycus}) fork lengths from a subsample of fish captured near Grand Haven, Lake Michigan, 1977.
}
\details{
A random subsample of the lengths in the \code{TroutperchLM2} data frame.
}
\section{Topic(s)}{

 \itemize{
   \item Length Frequency
   \item Length Expansion
   \item Size Structure
 }
}
\examples{
data(TroutperchLM3)
str(TroutperchLM3)
head(TroutperchLM3)
}
\concept{
'Length Frequency' 'Size Structure' 'Length Expansion'
}
\keyword{datasets}

