% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhitefishLS.R
\docType{data}
\name{WhitefishLS}
\alias{WhitefishLS}
\title{Landings and value of Lake Superior Lake Whitefish.}
\format{
A data frame with 12 observations on the following 4 variables:
 \describe{
   \item{year}{Year of catch.}
   \item{state}{State of catch (\code{MI}, \code{MN}, or \code{WI}).}
   \item{catch}{Catch in lbs.}
   \item{value}{Value of catch in dollars.}
 }
}
\source{
Results from a query in 2006 to the NMFS Great Lakes Commercial Fishery Landings database. \href{https://raw.githubusercontent.com/fishR-Core-Team/FSAdata/main/data-raw/WhitefishLS.csv}{CSV file}
}
\description{
Landings and value of Lake Whitefish \emph{Coregonus clupeaformis} in Lake Superior by state from 1971-2001.
}
\section{Topic(s)}{

 \itemize{
   \item Other
 }
}

\examples{
data(WhitefishLS)
str(WhitefishLS)
head(WhitefishLS)

}
\concept{Other}
\keyword{datasets}
