% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiteGrunt1.R
\docType{data}
\name{WhiteGrunt1}
\alias{WhiteGrunt1}
\title{Catch-at-age for White Grunt.}
\format{
A data frame with 25 observations on the following 2 variables.
 \describe{
   \item{age}{A numeric vector of assigned ages (from otoliths).}
   \item{catch}{A numeric vector of number of fish.} 
 }
}
\source{
From Figure 7 of Araujo, J.N. and A.S. Martins. 2007. Age, growth and mortality of white grunt (\emph{Haemulon plumierii}) from the central coast of Brazil. Scientia Marina 71:793-800. \href{https://raw.githubusercontent.com/fishR-Core-Team/FSAdata/main/data-raw/WhiteGrunt1.csv}{CSV file}
}
\description{
Catch-at-age for White Grunt (\emph{Haemulon plumierii}) collected from the central coast of Brazil.
}
\section{Topic(s)}{

 \itemize{
   \item Mortality 
   \item Catch curve 
 }
}

\examples{
data(WhiteGrunt1)
str(WhiteGrunt1)
head(WhiteGrunt1)
plot(log(catch)~age,data=WhiteGrunt1)

}
\concept{Catch Curve}
\concept{Mortality}
\keyword{datasets}
